+++ include/ctl/ccutl.h
@@ -725,12 +725,12 @@ namespace CCUTL_NAMESPACE {
 /**
  * Raises 10 to an unsigned intgral power.
  *
- * @tparam Res resulting unsigned_integral type
+ * @tparam Res resulting integral type
  * @tparam Exp unsigned_integral exponent
  * @param  n power to raise
  * @return 10^n
  */
-template <std::unsigned_integral Res, std::unsigned_integral Exp>
+template <std::integral Res, std::unsigned_integral Exp>
 [[nodiscard]] inline constexpr Res
 pow10(Exp n) noexcept {
   Res res = 1;
@@ -742,7 +742,7 @@ pow10(Exp n) noexcept {
 /**
  * Raises 10 to an unsigned intgral power.
  *
- * @tparam Res resulting unsigned_integral type
+ * @tparam Res resulting unsigned_integral type and exponent
  * @param  n power to raise
  * @return 10^n
  */